Changed bitmapMaxSize input to use a slider
Range is maxTextureSize/2 .. maxTextureSize, set to 5 steps in SettingsAdvancedScreen

+++ app/src/main/java/eu/kanade/tachiyomi/data/coil/TachiyomiImageDecoder.kt
@@ -30,7 +30,7 @@ class TachiyomiImageDecoder(private val resources: ImageSource, private val opti
 
         check(decoder != null && decoder.width > 0 && decoder.height > 0) { "Failed to initialize decoder" }
 
-        val maxBitmapSize by lazy { Injekt.get<BasePreferences>().maxBitmapSize().get().toInt() }
+        val maxBitmapSize by lazy { Injekt.get<BasePreferences>().maxBitmapSize().get() }
 
         val srcWidth = decoder.width
         val srcHeight = decoder.height
